+++ src/Trajectory/Trajectory.jl
@@ -1,4 +1,4 @@
-module Trajectory
+@reexport module Trajectory
 
 include("empirical_process.jl")
 include("generate_timeseries.jl")
