+++ src/TransitionMatrix/TransitionMatrix.jl
@@ -1,6 +1,5 @@
-module TransitionMatrix
+@reexport module TransitionMatrix
 
 include("construct_from_data.jl")
-include("utils.jl")
 
 end # module TransitionMatrix
